fix(api): let an explicit warranty refresh override the virtual-machine skip (#3201)
