+++ apps/api/src/services/warrantySync.test.ts
@@ -3,21 +3,34 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 // Verify upsertAgentWarranty maps agent-reported coverage kind to the right
 // persisted status / is_subscription flag without a live DB (#1320).
 const insertMock = vi.fn();
+const selectMock = vi.fn();
 
 vi.mock('../db', () => ({
   db: {
     insert: (...args: unknown[]) => insertMock(...args),
+    select: (...args: unknown[]) => selectMock(...args),
   },
 }));
 
 vi.mock('../db/schema', () => ({
   deviceWarranty: { deviceId: 'deviceWarranty.deviceId' },
-  deviceHardware: {},
-  devices: {},
+  deviceHardware: {
+    deviceId: 'deviceHardware.deviceId',
+    serialNumber: 'deviceHardware.serialNumber',
+    manufacturer: 'deviceHardware.manufacturer',
+    model: 'deviceHardware.model',
+  },
+  devices: {
+    id: 'devices.id',
+    orgId: 'devices.orgId',
+    isEphemeral: 'devices.isEphemeral',
+    isVirtual: 'devices.isVirtual',
+  },
 }));
 
+const providerMock = vi.fn();
 vi.mock('./warrantyProviders', () => ({
-  getProviderForManufacturer: vi.fn(),
+  getProviderForManufacturer: (...args: unknown[]) => providerMock(...args),
   normalizeManufacturer: (m: string) => m.toLowerCase(),
 }));
 
@@ -26,7 +39,7 @@ vi.mock('./warrantyAlertEvaluator', () => ({
   evaluateWarrantyAlerts: (...args: unknown[]) => evaluateWarrantyAlertsMock(...args),
 }));
 
-import { upsertAgentWarranty } from './warrantySync';
+import { syncWarrantyForDevice, upsertAgentWarranty } from './warrantySync';
 
 const DEVICE_ID = '44444444-4444-4444-4444-444444444444';
 const ORG_ID = '11111111-1111-1111-1111-111111111111';
@@ -119,5 +132,66 @@ describe('upsertAgentWarranty coverage-kind mapping', () => {
     expect(values).toHaveBeenCalledWith(
       expect.objectContaining({ status: 'expiring', isSubscription: false })
     );
+  });
+});
+
+describe('syncWarrantyForDevice — virtual machine exclusion (#3201)', () => {
+  // Reuses the file-level insertMock/db mock; adds a select chain, since the
+  // sync path reads hardware then the device row before touching a provider.
+  function queueReads(...results: unknown[][]) {
+    const queue = [...results];
+    selectMock.mockImplementation(() => ({
+      from: () => ({ where: () => ({ limit: () => Promise.resolve(queue.shift() ?? []) }) }),
+    }));
+  }
+
+  const VM_HARDWARE = [{
+    serialNumber: 'VMware-42 0f 1a',
+    manufacturer: 'VMware, Inc.',
+    model: 'VMware Virtual Platform',
+  }];
+  const PHYSICAL_HARDWARE = [{
+    serialNumber: 'ABC1234',
+    manufacturer: 'Dell Inc.',
+    model: 'Latitude 7420',
+  }];
+
+  beforeEach(() => {
+    selectMock.mockReset();
+    providerMock.mockReset();
+    providerMock.mockReturnValue(undefined);
+    insertMock.mockReturnValue({
+      values: () => ({ onConflictDoUpdate: () => Promise.resolve(undefined) }),
+    });
+  });
+
+  it('never consults a vendor provider for a virtual machine', async () => {
+    queueReads(VM_HARDWARE, [{ orgId: ORG_ID, isEphemeral: false, isVirtual: true }]);
+    await syncWarrantyForDevice(DEVICE_ID);
+    expect(providerMock).not.toHaveBeenCalled();
+  });
+
+  it('still consults a provider for a physical device', async () => {
+    queueReads(PHYSICAL_HARDWARE, [{ orgId: ORG_ID, isEphemeral: false, isVirtual: false }]);
+    await syncWarrantyForDevice(DEVICE_ID);
+    expect(providerMock).toHaveBeenCalledWith('Dell Inc.');
+  });
+
+  it('still excludes ephemeral Quick Support devices', async () => {
+    queueReads(PHYSICAL_HARDWARE, [{ orgId: ORG_ID, isEphemeral: true, isVirtual: false }]);
+    await syncWarrantyForDevice(DEVICE_ID);
+    expect(providerMock).not.toHaveBeenCalled();
+  });
+
+  it('force: an explicit user refresh DOES sync a virtual machine', async () => {
+    queueReads(VM_HARDWARE, [{ orgId: ORG_ID, isEphemeral: false, isVirtual: true }]);
+    await syncWarrantyForDevice(DEVICE_ID, { force: true });
+    expect(providerMock).toHaveBeenCalledWith('VMware, Inc.');
+  });
+
+  it('force does NOT bypass the ephemeral skip — that is an ownership rule', async () => {
+    queueReads(PHYSICAL_HARDWARE, [{ orgId: ORG_ID, isEphemeral: true, isVirtual: false }]);
+    await syncWarrantyForDevice(DEVICE_ID, { force: true });
+    expect(providerMock).not.toHaveBeenCalled();
   });
 });
